requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
urllib3==1.26.6
pymongo==3.12.0
requests==2.26.0
selenium==3.141.0
python-dotenv==0.19.0
dnspython==2.1.0
